update"join" listener to consider link join
as opposed to normal join by id from AppRoom or joinRandom

this is because when user joined from creating the room (GameEmpty),
or joinGame, user also calls an api call that update the DB information
of that game. Meanwhile, the user that joined from immediately entering
the game link haven't updated the Game in DB yet, so

in socket server, "join" listener, socket will also check the Game in
db if the uid that the user sent is inside the DB or not. IOW, check if
user1 is still empty and user0 is not uid.
if true, socket will update the DB for the user.