add "online" prop for player, and remodel joinUser
when the game starts and beyond (game "playing" or "ended"), we don't
change the value of user0 and user1 anymore. They are locked for the
rest of their life. Instead, we now look at pwhite and pblack (players)
and modify their "online" props to true/false depending on if they
join/leave during the game