refactor repetitive model's findById and check
create getUser, getChat, getGame in utils.js that does what every
other routes & socket is doing when searching for a document:
get by id, then check if it is null. If null, throw a not found error.

this greatly simplify the codebase! (by a bit, just couples of 1 lines
after all)